# book-social-api/src/main/java/com/sparklecow/book/dto/user/UserResponseDto.java
package com.sparklecow.book.dto.user;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserResponseDto{
    private String token;
}